CUDASetup only executed once + fixed circular import.
