--- bitsandbytes/cuda_setup/__init__.py
@@ -1,6 +1,0 @@
-from .main import evaluate_cuda_setup
-from .paths import (
-    CUDA_RUNTIME_LIB,
-    determine_cuda_runtime_lib_path,
-    extract_candidate_paths,
-)
